Doc comments, deduplicate code
