+++ src/Microsoft.Extensions.Logging.Configuration/LoggerProviderOptionsChangeTokenSource.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.Extensions.Options;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.Extensions.Options;
 
 namespace Microsoft.Extensions.Logging
 {
